Move copy_recursively to utils
