+++ buildpacks/dotnet/src/utils.rs
@@ -1,5 +1,6 @@
-use std::io;
+use std::path::Path;
 use std::process::{Command, ExitStatus};
+use std::{fs, io};
 
 /// A helper for running an external process using [`Command`], that streams stdout/stderr
 /// to the user and checks that the exit status of the process was non-zero.
@@ -16,6 +17,22 @@ pub(crate) fn run_command_and_stream_output(
                 Err(StreamedCommandError::NonZeroExitStatus(exit_status))
             }
         })
+}
+
+pub(crate) fn copy_recursively<P: AsRef<Path>>(src: P, dst: P) -> std::io::Result<()> {
+    if src.as_ref().is_dir() {
+        fs::create_dir_all(dst.as_ref())?;
+        for entry in fs::read_dir(src)? {
+            let entry = entry?;
+            let src_path = entry.path();
+            let dst_path = dst.as_ref().join(entry.file_name());
+
+            copy_recursively(&src_path, &dst_path)?;
+        }
+    } else {
+        fs::copy(src, dst)?;
+    }
+    Ok(())
 }
 
 /// Errors that can occur when running an external process using `run_command_and_stream_output`.
